src/main/java/com/almondtools/comtemplate/engine/expressions/EvalContextVar.java
package com.almondtools.comtemplate.engine.expressions;

import com.almondtools.comtemplate.engine.Scope;
import com.almondtools.comtemplate.engine.TemplateExpression;
import com.almondtools.comtemplate.engine.TemplateExpressionVisitor;

public class EvalContextVar implements TemplateExpression {

	private String name;

	public EvalContextVar(String name) {
		this.name = name;
	}
	
	public String getName() {
		return name;
	}
	
	@Override
	public <T> T apply(TemplateExpressionVisitor<T> visitor, Scope scope) {
		return visitor.visitEvalContextVar(this, scope);
	}

	@Override
	public String toString() {
		return "@@" + name;
	}
}
